+++ config.h
@@ -436,6 +436,7 @@ struct config {
         bool sixel;
         enum shm_bit_depth surface_bit_depth;
         uint32_t min_stride_alignment;
+        bool preapply_damage;
     } tweak;
 
     struct {